requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0

# LLM and embeddings
openai==1.3.0
langchain==0.0.335
langchain-openai==0.0.2
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.1.1

# Vector databases
chromadb==0.4.18
qdrant-client==1.6.4
pymilvus==2.3.1

# Text processing
nltk==3.8.1
spacy==3.7.2
beautifulsoup4==4.12.2
lxml==4.9.3
markdown==3.5.1
python-docx==1.0.1
PyPDF2==3.0.1

# Search and retrieval
rank-bm25==0.2.2
faiss-cpu==1.7.4
scikit-learn==1.3.2
numpy==1.26.1
pandas==2.1.3

# API and web
httpx==0.25.1
aiohttp==3.8.6
jinja2==3.1.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utilities
tqdm==4.66.1
loguru==0.7.2
tenacity==8.2.3 